Allow Response to assume only replies from Owner and AdminCCs meet SLA
By default, a reply by anyone who is not a requestor unsets the due date
for a Response SLA.  If it is common for non-requestors to reply to
tickets, this isn't desirable.  This commit provides a way to only treat
replies from Owners and AdminCcs as meeting a Response SLA.

IsRequestorsAct is renamed to IsOutsideActor, and the documentation for
Response SLAs is updated to clarify what is meant by an "outside actor."
The new behavior is controlled by a new AssumeOutsideActor global SLA
configuration option.